fixes for seaportcreatelisting examples, moving examples to appropriate folder
